refactor: Move `doStorybookHydration` closer to its friends
