+++ dotcom-rendering/src/web/layouts/InteractiveImmersive.stories.tsx
@@ -16,7 +16,7 @@ import { injectPrivacySettingsLink } from '@root/src/web/lib/injectPrivacySettin
 import { extractNAV } from '@root/src/model/extract-nav';
 import { fireAndResetHydrationState } from '@root/src/web/components/HydrateOnce';
 import { DecideLayout } from './DecideLayout';
-import { doStorybookHydration } from '../lib/doStorybookHydration';
+import { doStorybookHydration } from '../browser/islands/doStorybookHydration';
 
 mockRESTCalls();
