+++ dotcom-rendering/src/web/layouts/Standard.stories.tsx
@@ -34,7 +34,7 @@ import { injectPrivacySettingsLink } from '@root/src/web/lib/injectPrivacySettin
 import { extractNAV } from '@root/src/model/extract-nav';
 import { fireAndResetHydrationState } from '@root/src/web/components/HydrateOnce';
 import { DecideLayout } from './DecideLayout';
-import { doStorybookHydration } from '../lib/doStorybookHydration';
+import { doStorybookHydration } from '../browser/islands/doStorybookHydration';
 
 mockRESTCalls();
